Make `Chart.layout(Service)` importable
Rename (and deprecate) `Chart.layoutService` to `Chart.layout` and make it importable.
