+++ src/plugins/plugin.legend.js
@@ -3,6 +3,7 @@
 var defaults = require('../core/core.defaults');
 var Element = require('../core/core.element');
 var helpers = require('../helpers/index');
+var layout = require('../core/core.layout');
 
 defaults._set('global', {
 	legend: {
@@ -81,7 +82,6 @@ defaults._set('global', {
 
 module.exports = function(Chart) {
 
-	var layout = Chart.layoutService;
 	var noop = helpers.noop;
 
 	/**
